% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allThinnedInts.R
\name{all.thinned.ints}
\alias{all.thinned.ints}
\title{all.thinned.ints
Takes a list of MNIST digits, removes non-white pixels and
calls thin.points,the thinning algorithm for creating a pixel "skeleton"}
\usage{
\method{all}{thinned.ints}(digits, limit = NA)
}
\arguments{
\item{digits:}{list of MNIST digits to thin}

\item{limit:}{optional; in case entire list of MNIST digits is passed, can limit number passed to thinning algorithm}
}
\value{
thinned digits, pixel "skeletons"
}
\description{
all.thinned.ints
Takes a list of MNIST digits, removes non-white pixels and
calls thin.points,the thinning algorithm for creating a pixel "skeleton"
}

