% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectDigitSkeletons.R
\name{collect.digit.skeletons}
\alias{collect.digit.skeletons}
\title{collect.digit.skeletons
Takes a list of MNIST digits, removes non-white pixels and
calls thin.points,the thinning algorithm for creating a pixel "skeleton"}
\usage{
collect.digit.skeletons(digits, limit = NA)
}
\arguments{
\item{digits:}{list of MNIST digits to thin}

\item{limit:}{optional; in case entire list of MNIST digits is passed, can limit number passed to thinning algorithm}
}
\value{
thinned digits, pixel "skeletons"
}
\description{
collect.digit.skeletons
Takes a list of MNIST digits, removes non-white pixels and
calls thin.points,the thinning algorithm for creating a pixel "skeleton"
}
\examples{
# coming soon
}

