% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectTraining.R
\name{collect.training.set}
\alias{collect.training.set}
\title{Title}
\usage{
collect.training.set(thinned.ints, set.length = 10)
}
\arguments{
\item{thinned.ints}{}
}
\value{
# coming soon
}
\description{
Title
}
\examples{
# coming soon
}

