% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/displayDigit.R
\name{displayDigit}
\alias{displayDigit}
\title{displayDigit
Author: Hrvoje Stojic
Prints a 16x16 MNIST digit using pixel values}
\usage{
displayDigit(pixels, label, predictedLabel = NULL, saveImage = FALSE,
  newDevice = TRUE)
}
\arguments{
\item{pixels}{}

\item{label}{}

\item{predictedLabel}{}

\item{saveImage}{}

\item{newDevice}{}
}
\value{
# coming soon
}
\description{
displayDigit
Author: Hrvoje Stojic
Prints a 16x16 MNIST digit using pixel values
}
\examples{
# coming soon
}

