% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generatePaths.R
\name{generatePaths}
\alias{generatePaths}
\title{generatePaths
Accepts a list of thinned ints (digit skeletons) and returns a list of thinned ints' characteristics (see value)}
\usage{
generatePaths(thinned.ints, animation = FALSE, sleep.time = 0.2)
}
\arguments{
\item{thinned.ints}{}

\item{animation}{optional, whether path generation algorithm}

\item{sleep.time}{}
}
\value{
List of each thinned int's characteristics as determined by the algorithm:
  number of strokes, path of points followed, direction steps (vector of 1-8 representing 8 directions possible from any one cell),
  changes in direction, and loops
  Used to make prediction on test sets.
}
\description{
generatePaths
Accepts a list of thinned ints (digit skeletons) and returns a list of thinned ints' characteristics (see value)
}
\examples{
# coming soon
}

