% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nearest.point}
\alias{nearest.point}
\title{Title}
\usage{
nearest.point(point, points, order = 1)
}
\arguments{
\item{point}{}

\item{points}{}

\item{order}{}
}
\value{
# coming soon
}
\description{
Title
}
\examples{
# coming soon
}

