% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nearorfar.from.origin}
\alias{nearorfar.from.origin}
\title{Title}
\usage{
nearorfar.from.origin(points, relativity = "nearest")
}
\arguments{
\item{points}{}

\item{relativity}{}
}
\value{
# coming soon
}
\description{
Title
}
\examples{
# coming soon
}

