% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot.point}
\alias{plot.point}
\title{plot.point
color the point on the plot with given color}
\usage{
\method{plot}{point}(point, color = "cadetblue")
}
\arguments{
\item{point}{}

\item{color}{defaults to cadetblue}
}
\description{
plot.point
color the point on the plot with given color
}

