% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictMnist.r
\name{predict.mnist}
\alias{predict.mnist}
\title{Title}
\usage{
\method{predict}{mnist}(train.objects, test.set)
}
\arguments{
\item{train.objects}{}

\item{test.set}{}
}
\value{
# coming soon
}
\description{
Title
}
\examples{
# coming soon
}

