% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relativePosition.R
\name{relative.position}
\alias{relative.position}
\title{relative.position}
\usage{
relative.position(point1, point2)
}
\arguments{
\item{point1}{}

\item{point2}{}
}
\value{
position 1-8 (where 1 is top left corner and increments are made in clockwise direction) of point2 relative to point1
}
\description{
relative.position
}

