% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thinPoints.R
\name{thin.points}
\alias{thin.points}
\title{thin.points
Takes a starting point and pixels and returns a thinned version / pixel skeleton}
\usage{
thin.points(current.point, white.pixels, animation = FALSE)
}
\arguments{
\item{current.point}{}

\item{white.pixels}{}

\item{animation}{}
}
\value{
# coming soon
}
\description{
thin.points
Takes a starting point and pixels and returns a thinned version / pixel skeleton
}
\examples{
# coming soon
}

