% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainDigit.R
\name{train.digit}
\alias{train.digit}
\title{train.digit
takes a thinned digit, generates a directed path for that digit
  and returns a vector of probabilities for each step type, the number of steps in the 
  and the label of the digit}
\usage{
train.digit(digit, animation = FALSE)
}
\arguments{
\item{digit}{}
}
\description{
train.digit
takes a thinned digit, generates a directed path for that digit
  and returns a vector of probabilities for each step type, the number of steps in the 
  and the label of the digit
}

